requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
requests==2.31.0
opencv-python==4.8.1.78
RPi.GPIO==0.7.1
smbus2==0.4.3
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0
schedule==1.2.0
croniter==1.4.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
marshmallow==3.20.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0



